Abstract tasks for different deployments
