+++ gwvolman/tasks_kubernetes.py
@@ -1,0 +1,4 @@
+from .tasks_base import TasksBase
+
+class KubernetesTasks(TasksBase):
+    pass
